requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
psycopg2-binary==2.7.4
python-crontab==2.2.8
requests==2.18.4
six==1.11.0
sqlalchemy==1.2.7
testing.common.database==2.0.3
testing.postgresql==1.3.0
urllib3==1.22
Werkzeug==0.12.2
